pyproject.toml
[project]
name = "swjtu-score-monitor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "azure-functions>=1.24.0",
    "bs4==0.0.2",
    "ddddocr==1.5.6",
    "fastapi==0.99.1",
    "psycopg2-binary==2.9.11",
    "requests==2.32.5",
]
